improv: history ui
